Better UI for FormError
Add the FormErrors in the ActionUiState and kept the Errors. The difference is that the FormErrors should be used when there is a error in the FormTextField (eg. business rule). The Errors should be used for errors happening not related to forms (eg. internal database error). The FormTextField component was updated accordingly.